Module aws_sdk_amp::client::fluent_builders

Utilities to ergonomically construct a request to the service.

Fluent builders are created through the Client by calling one if its operation methods. After parameters are set using the builder methods, the send method can be called to initiate the request.

Structs

CreateAlertManagerDefinition

Fluent builder constructing a request to CreateAlertManagerDefinition.

CreateRuleGroupsNamespace

Fluent builder constructing a request to CreateRuleGroupsNamespace.

CreateWorkspace

Fluent builder constructing a request to CreateWorkspace.

DeleteAlertManagerDefinition

Fluent builder constructing a request to DeleteAlertManagerDefinition.

DeleteRuleGroupsNamespace

Fluent builder constructing a request to DeleteRuleGroupsNamespace.

DeleteWorkspace

Fluent builder constructing a request to DeleteWorkspace.

DescribeAlertManagerDefinition

Fluent builder constructing a request to DescribeAlertManagerDefinition.

DescribeRuleGroupsNamespace

Fluent builder constructing a request to DescribeRuleGroupsNamespace.

DescribeWorkspace

Fluent builder constructing a request to DescribeWorkspace.

ListRuleGroupsNamespaces

Fluent builder constructing a request to ListRuleGroupsNamespaces.

ListTagsForResource

Fluent builder constructing a request to ListTagsForResource.

ListWorkspaces

Fluent builder constructing a request to ListWorkspaces.

PutAlertManagerDefinition

Fluent builder constructing a request to PutAlertManagerDefinition.

PutRuleGroupsNamespace

Fluent builder constructing a request to PutRuleGroupsNamespace.

TagResource

Fluent builder constructing a request to TagResource.

UntagResource

Fluent builder constructing a request to UntagResource.

UpdateWorkspaceAlias

Fluent builder constructing a request to UpdateWorkspaceAlias.