Module aws_sdk_amp::error

Errors that can occur when calling the service.

Modules

access_denied_exception

See AccessDeniedException

conflict_exception

See ConflictException

internal_server_exception

See InternalServerException

resource_not_found_exception

See ResourceNotFoundException

service_quota_exceeded_exception

See ServiceQuotaExceededException

throttling_exception

See ThrottlingException

validation_exception

See ValidationException

Structs

AccessDeniedException

User does not have sufficient access to perform this action.

ConflictException

Updating or deleting a resource can cause an inconsistent state.

CreateAlertManagerDefinitionError

Error type for the CreateAlertManagerDefinition operation.

CreateRuleGroupsNamespaceError

Error type for the CreateRuleGroupsNamespace operation.

CreateWorkspaceError

Error type for the CreateWorkspace operation.

DeleteAlertManagerDefinitionError

Error type for the DeleteAlertManagerDefinition operation.

DeleteRuleGroupsNamespaceError

Error type for the DeleteRuleGroupsNamespace operation.

DeleteWorkspaceError

Error type for the DeleteWorkspace operation.

DescribeAlertManagerDefinitionError

Error type for the DescribeAlertManagerDefinition operation.

DescribeRuleGroupsNamespaceError

Error type for the DescribeRuleGroupsNamespace operation.

DescribeWorkspaceError

Error type for the DescribeWorkspace operation.

InternalServerException

Unexpected error during processing of request.

ListRuleGroupsNamespacesError

Error type for the ListRuleGroupsNamespaces operation.

ListTagsForResourceError

Error type for the ListTagsForResource operation.

ListWorkspacesError

Error type for the ListWorkspaces operation.

PutAlertManagerDefinitionError

Error type for the PutAlertManagerDefinition operation.

PutRuleGroupsNamespaceError

Error type for the PutRuleGroupsNamespace operation.

ResourceNotFoundException

Request references a resource which does not exist.

ServiceQuotaExceededException

Request would cause a service quota to be exceeded.

TagResourceError

Error type for the TagResource operation.

ThrottlingException

Request was denied due to request throttling.

UntagResourceError

Error type for the UntagResource operation.

UpdateWorkspaceAliasError

Error type for the UpdateWorkspaceAlias operation.

ValidationException

The input fails to satisfy the constraints specified by an AWS service.

Enums

CreateAlertManagerDefinitionErrorKind

Types of errors that can occur for the CreateAlertManagerDefinition operation.

CreateRuleGroupsNamespaceErrorKind

Types of errors that can occur for the CreateRuleGroupsNamespace operation.

CreateWorkspaceErrorKind

Types of errors that can occur for the CreateWorkspace operation.

DeleteAlertManagerDefinitionErrorKind

Types of errors that can occur for the DeleteAlertManagerDefinition operation.

DeleteRuleGroupsNamespaceErrorKind

Types of errors that can occur for the DeleteRuleGroupsNamespace operation.

DeleteWorkspaceErrorKind

Types of errors that can occur for the DeleteWorkspace operation.

DescribeAlertManagerDefinitionErrorKind

Types of errors that can occur for the DescribeAlertManagerDefinition operation.

DescribeRuleGroupsNamespaceErrorKind

Types of errors that can occur for the DescribeRuleGroupsNamespace operation.

DescribeWorkspaceErrorKind

Types of errors that can occur for the DescribeWorkspace operation.

ListRuleGroupsNamespacesErrorKind

Types of errors that can occur for the ListRuleGroupsNamespaces operation.

ListTagsForResourceErrorKind

Types of errors that can occur for the ListTagsForResource operation.

ListWorkspacesErrorKind

Types of errors that can occur for the ListWorkspaces operation.

PutAlertManagerDefinitionErrorKind

Types of errors that can occur for the PutAlertManagerDefinition operation.

PutRuleGroupsNamespaceErrorKind

Types of errors that can occur for the PutRuleGroupsNamespace operation.

TagResourceErrorKind

Types of errors that can occur for the TagResource operation.

UntagResourceErrorKind

Types of errors that can occur for the UntagResource operation.

UpdateWorkspaceAliasErrorKind

Types of errors that can occur for the UpdateWorkspaceAlias operation.